Moved the godep strip functionality to a custom package
